# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: ShelterFlex Backend API
  description: API for ShelterFlex backend service
  version: 0.1.0
  contact:
    name: ShelterFlex Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.shelterflex.io
    description: Production server
paths:
...
  /api/deals:
    post:
      summary: Create a new deal with repayment schedule
      description: Creates a new financing deal with a deterministic repayment schedule
      operationId: createDeal
      tags:
        - Deals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealRequest'
      responses:
        '201':
          description: Deal created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DealWithSchedule'
        '400':
          description: Validation error or listing not approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                validationError:
                  summary: Invalid request data
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid request data
                listingNotApproved:
                  summary: Listing not approved
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Listing must be approved to create a deal. Current status pending_review
        '404':
          description: Listing not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: NOT_FOUND
                  message: Listing with ID '550e8400-e29b-41d4-a716-446655440000' not found
        '409':
          description: Listing already rented or linked to another deal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                listingRented:
                  summary: Listing already rented
                  value:
                    error:
                      code: LISTING_ALREADY_RENTED
                      message: Listing with ID '550e8400-e29b-41d4-a716-446655440000' is already rented
                listingLinked:
                  summary: Listing already linked to a deal
                  value:
                    error:
                      code: LISTING_ALREADY_RENTED
                      message: Listing with ID '550e8400-e29b-41d4-a716-446655440000' is already linked to deal 550e8400-e29b-41d4-a716-446655440001
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'

    get:
      summary: Get deals with optional filtering
      description: Retrieves a paginated list of deals, optionally filtered by tenant, landlord, or status
      operationId: getDeals
      tags:
        - Deals
      parameters:
        - name: tenantId
          in: query
          description: Filter by tenant ID
          schema:
            type: string
        - name: landlordId
          in: query
          description: Filter by landlord ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by deal status
          schema:
            $ref: '#/components/schemas/DealStatus'
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Deals retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/PaginatedDeals'
        '400':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /api/deals/{dealId}:
    get:
      summary: Get a specific deal by ID with schedule
      description: Retrieves a single deal including its full repayment schedule
      operationId: getDealById
      tags:
        - Deals
      parameters:
        - name: dealId
          in: path
          required: true
          description: ID of the deal to retrieve
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deal retrieved successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DealWithSchedule'
        '404':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /api/deals/{dealId}/status:
    patch:
      summary: Update deal status
      description: Updates the status of a deal (e.g., from draft to active)
      operationId: updateDealStatus
      tags:
        - Deals
      parameters:
        - name: dealId
          in: path
          required: true
          description: ID of the deal to update
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDealStatusRequest'
      responses:
        '200':
          description: Deal status updated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DealWithSchedule'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /api/deals/{dealId}/schedule/{period}:
    patch:
      summary: Update schedule item status
      description: Updates the status of a specific repayment schedule item
      operationId: updateScheduleItemStatus
      tags:
        - Deals
      parameters:
        - name: dealId
          in: path
          required: true
          description: ID of the deal
          schema:
            type: string
            format: uuid
        - name: period
          in: path
          required: true
          description: Period number in the repayment schedule
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleItemRequest'
      responses:
        '200':
          description: Schedule item status updated successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DealWithSchedule'
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
  /api/deals/{dealId}/progress:
    get:
      summary: Get deal payment progress from on-chain receipts
      description: |
        Computes and returns a deal's payment progress by reading `TENANT_REPAYMENT`
        receipts from the on-chain ledger (outbox store).

        Only receipts with status `sent` (confirmed on-chain) are counted.
        USDC is the canonical accounting unit.
      operationId: getDealProgress
      tags:
        - Deals
      parameters:
        - name: dealId
          in: path
          required: true
          description: ID of the deal
          schema:
            type: string
            format: uuid
            example: 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Deal progress computed from on-chain receipts
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/DealProgressResponse'
              example:
                success: true
                data:
                  totalPaidUsdc: "129.040000"
                  periodsPaid: 2
                  remainingPeriods: 10
                  nextDueDate: "2024-04-15T00:00:00.000Z"
                  lastPaymentTxId: "a3f2c1d4e5b6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2"
                  lastPaymentExternalRefSource: stripe
                  lastPaymentExternalRef: pi_3OZxyz123456
        '404':
          description: Deal not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error:
                  code: NOT_FOUND
                  message: Deal with ID '550e8400-e29b-41d4-a716-446655440000' not found
        '429':
          $ref: '#/components/responses/TooManyRequests'
        default:
          $ref: '#/components/responses/Error'
...
components:
  schemas:
...
    DealStatus:
      type: string
      enum:
        - draft
        - active
        - completed
        - defaulted
      example: draft
    ScheduleItemStatus:
      type: string
      enum:
        - upcoming
        - due
        - paid
        - late
      example: upcoming
    ScheduleItem:
      type: object
      required:
        - period
        - dueDate
        - amountNgn
        - status
      properties:
        period:
          type: integer
          description: Period number in the repayment schedule (1-indexed)
          example: 1
        dueDate:
          type: string
          format: date-time
          description: ISO date when payment is due
          example: 2024-02-15T00:00:00Z
        amountNgn:
          type: number
          format: double
          description: Payment amount in Nigerian Naira
          example: 83333.33
        status:
          $ref: '#/components/schemas/ScheduleItemStatus'
    Deal:
      type: object
      required:
        - dealId
        - tenantId
        - landlordId
        - annualRentNgn
        - depositNgn
        - financedAmountNgn
        - termMonths
        - createdAt
        - status
      properties:
        dealId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        tenantId:
          type: string
          description: ID of the tenant repaying the financing
          example: tenant-001
        landlordId:
          type: string
          description: ID of the landlord who received upfront payment
          example: landlord-001
        listingId:
          type: string
          format: uuid
          nullable: true
          description: Optional ID of the associated property listing
          example: 550e8400-e29b-41d4-a716-446655440001
        annualRentNgn:
          type: integer
          description: Annual rent amount in Naira
          example: 1200000
        depositNgn:
          type: integer
          description: Deposit amount paid by tenant (must be >= 20% of annual rent)
          example: 240000
        financedAmountNgn:
          type: integer
          description: Amount financed by ShelterFlex (annual rent - deposit)
          example: 960000
        termMonths:
          type: integer
          description: Repayment term in months (3, 6, or 12)
          example: 12
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:00:00Z
        status:
          $ref: '#/components/schemas/DealStatus'
    DealWithSchedule:
      type: object
      required:
        - dealId
        - tenantId
        - landlordId
        - annualRentNgn
        - depositNgn
        - financedAmountNgn
        - termMonths
        - createdAt
        - status
        - schedule
      properties:
        dealId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        tenantId:
          type: string
          example: tenant-001
        landlordId:
          type: string
          example: landlord-001
        listingId:
          type: string
          format: uuid
          nullable: true
          example: 550e8400-e29b-41d4-a716-446655440001
        annualRentNgn:
          type: integer
          example: 1200000
        depositNgn:
          type: integer
          example: 240000
        financedAmountNgn:
          type: integer
          example: 960000
        termMonths:
          type: integer
          example: 12
        createdAt:
          type: string
          format: date-time
          example: 2024-01-15T10:00:00Z
        status:
          $ref: '#/components/schemas/DealStatus'
        schedule:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleItem'
    CreateDealRequest:
      type: object
      required:
        - tenantId
        - landlordId
        - annualRentNgn
        - depositNgn
        - termMonths
      properties:
        tenantId:
          type: string
          description: ID of the tenant who will repay the financing
          example: tenant-001
        landlordId:
          type: string
          description: ID of the landlord who will receive upfront payment
          example: landlord-001
        listingId:
          type: string
          format: uuid
          nullable: true
          description: Optional ID of the associated property listing
          example: 550e8400-e29b-41d4-a716-446655440001
        annualRentNgn:
          type: integer
          description: Annual rent amount in Naira
          example: 1200000
        depositNgn:
          type: integer
          description: Deposit amount paid by tenant (must be >= 20% of annual rent)
          example: 240000
        termMonths:
          type: integer
          description: Repayment term in months (must be 3, 6, or 12)
          example: 12
    UpdateDealStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/DealStatus'
    UpdateScheduleItemRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/ScheduleItemStatus'
    PaginatedDeals:
      type: object
      required:
        - deals
        - total
        - page
        - pageSize
        - totalPages
      properties:
        deals:
          type: array
          items:
            $ref: '#/components/schemas/Deal'
        total:
          type: integer
          description: Total number of deals matching the filter
          example: 42
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of deals per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 3
...
    DealProgressResponse:
      type: object
      required:
        - totalPaidUsdc
        - periodsPaid
        - remainingPeriods
        - nextDueDate
      properties:
        totalPaidUsdc:
          type: string
          description: Total USDC paid across all confirmed on-chain TENANT_REPAYMENT receipts (decimal string with 6 decimal places)
          example: "129.040000"
        periodsPaid:
          type: integer
          description: Number of TENANT_REPAYMENT receipts confirmed on-chain
          example: 2
        remainingPeriods:
          type: integer
          description: Number of remaining repayment periods (termMonths minus periodsPaid, clamped to >= 0)
          example: 10
        nextDueDate:
          type: string
          format: date-time
          nullable: true
          description: ISO 8601 due date of the next scheduled payment; null if fully paid
          example: "2024-04-15T00:00:00.000Z"
        lastPaymentTxId:
          type: string
          description: txId (SHA-256 hex) of the most recent confirmed receipt
          example: "a3f2c1d4e5b6f7a8b9c0d1e2f3a4b5c6d7e8f9a0b1c2d3e4f5a6b7c8d9e0f1a2"
        lastPaymentExternalRefSource:
          type: string
          description: Source part of the canonical external reference (e.g. "stripe", "manual")
          example: stripe
        lastPaymentExternalRef:
          type: string
          description: Reference part of the canonical external reference (e.g. payment intent ID)
          example: pi_3OZxyz123456
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - FORBIDDEN
                - TOO_MANY_REQUESTS
                - NOT_FOUND
                - CONFLICT
                - LISTING_ALREADY_RENTED
                - SOROBAN_ERROR
                - INTERNAL_ERROR
              example: NOT_FOUND
            message:
              type: string
              example: "Resource not found"
            details:
              type: object
              description: Optional structured context for the error
              additionalProperties: true
...
  responses:
    Error:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: "TOO_MANY_REQUESTS"
              message: "Rate limit exceeded"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
...